Refactor how help is displayed.
Using the game buffer isn't a good idea because the game table can be of
arbitrary length. If it's too tall, help will go off screen and the use has no
way of scrolling without pausing the game. Just use the same command in both
cases (during game and outside of game). This should be easier to maintain, too.